Added write encoding
